requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
groq
pypdf
sentence-transformers
scikit-learn
pandas
numpy
jinja2
python-dotenv
